refact: 컨트롤러 분리 및 타입설정
