+++ server/src/routes/index.ts
@@ -5,7 +5,7 @@ import * as posts from "./post.routes";
 import * as market from "./market.routes";
 import * as like from "./like.routes";
 import * as movepost from "./movepost.routes.js";
-// import * as search from "./search.routes.js";
+import * as search from "./search.routes.js";
 // import * as comment from "./comment.routes.js";
 
 import authmiddleware from "../util/authmiddleware";
@@ -20,7 +20,7 @@ router.use(posts.path, posts.router);
 router.use(market.path, market.router);
 router.use(like.path, authmiddleware, like.router);
 router.use(movepost.path, movepost.router);
-// router.use(search.path, search.router);
+router.use(search.path, search.router);
 // router.use(comment.path, authmiddleware, comment.router);
 
 export default router;
